pyproject.toml
[tool.poetry]
name = "gcpts"
version = "0.1.0"
description = ""
authors = ["richwomanbtc <richwomanbtc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.2"
gcsfs = "^2022.11.0"
pyarrow = "^10.0.1"
google-cloud-bigquery = "^3.4.1"
db-dtypes = "^1.0.5"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
flake8 = "^6.0.0"
black = {version = "^22.12.0", allow-prereleases = true}
pytest = "^7.2.0"
pytest-order = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
